refactor: decompose 5 more monolith files into submodules (issue #2486 sprint 3)
Continues sprint-based decomposition of files exceeding 500 LOC:

- terrain_engine.py (969 → 234 LOC) → _terrain_geometry, _terrain_physics, _terrain_utils
- myosuite_physics_engine.py (934 → 61 LOC) → _engine_init, _simulation_core,
  _dynamics, _muscle_interface, _drift_control
- coordination.py (972 → 29 LOC) → _coordination_alignment, _coordination_phase,
  _coordination_sequence, _coordination_synergy
- gas_properties.py (969 → 188 LOC) → _mixture_properties, _heat_capacity,
  _component_database, _viscosity
- _id_core.py further refined → _id_solver_mixin extracted (466 LOC)

All public symbols remain importable from original module paths.
