+++ src/engines/physics_engines/myosuite/python/myosuite_physics_engine.py
@@ -14,40 +14,26 @@ Refactored to use shared engine availability module (DRY principle).
 
 from __future__ import annotations
 
-from typing import Any
-
-# Path: src/engines/physics_engines/myosuite/python/myosuite_physics_engine.py -> need 6 parents
-import numpy as np  # noqa: E402
-
-from src.shared.python.core.contracts import (  # noqa: E402
-    check_finite,
-    postcondition,
-    precondition,
-)
-from src.shared.python.engine_core.engine_availability import (
-    MYOSUITE_AVAILABLE,  # noqa: E402
-)
 from src.shared.python.engine_core.interfaces import PhysicsEngine  # noqa: E402
 from src.shared.python.logging_pkg.logging_config import get_logger  # noqa: E402
+
+from ._drift_control import DriftControlMixin
+from ._dynamics import DynamicsMixin
+from ._engine_init import EngineInitMixin
+from ._muscle_interface import MuscleInterfaceMixin
+from ._simulation_core import SimulationCoreMixin
 
 logger = get_logger(__name__)
 
 
-# Log warning if MyoSuite not available
-
-if not MYOSUITE_AVAILABLE:
-    logger.warning("MyoSuite not installed. MyoSuitePhysicsEngine will not function.")
-
-else:
-    # Prefer gymnasium (modern fork) over legacy gym
-    try:
-        import gymnasium as gym
-    except ImportError:
-        import gym  # type: ignore[no-redef]
-    import myosuite  # noqa: F401
-
-
-class MyoSuitePhysicsEngine(PhysicsEngine):
+class MyoSuitePhysicsEngine(
+    EngineInitMixin,
+    SimulationCoreMixin,
+    DynamicsMixin,
+    DriftControlMixin,
+    MuscleInterfaceMixin,
+    PhysicsEngine,
+):
     """MyoSuite Engine Wrapper.
 
 
@@ -60,866 +46,7 @@ class MyoSuitePhysicsEngine(PhysicsEngine):
 
     def __init__(self) -> None:
         """Initialize."""
-
-        self.env: gym.Env | None = None
-
-        self.sim: Any = None  # Underlying mujoco simulation object
-
-        self.env_id: str = ""
-
-        self._dt = 0.002  # Default
-
-        self._terminated: bool = False  # True after episode termination or truncation
-
-    def _reset_loaded_state(self) -> None:
-        """Clear all engine state after a failed or closed load attempt."""
-        self.env = None
-        self.sim = None
-        self.env_id = ""
-        self._dt = 0.002
-
-    @staticmethod
-    def _extract_sim_from_env(env: Any) -> Any:
-        """Extract the underlying MuJoCo simulation object from a Gym env."""
-        sim = getattr(env, "sim", None)
-        if sim is not None:
-            return sim
-
-        unwrapped = getattr(env, "unwrapped", None)
-        if unwrapped is not None:
-            sim = getattr(unwrapped, "sim", None)
-            if sim is not None:
-                return sim
-
-        raise RuntimeError(
-            "Could not extract underlying MuJoCo sim object from MyoSuite env"
-        )
-
-    @property
-    def model_name(self) -> str:
-        """Return the Gym environment ID as the model name."""
-        return self.env_id or "MyoSuite_NoModel"
-
-    @property
-    def model(self) -> Any:  # mujoco.MjModel | None when available
-        """Expose underlying MuJoCo model for direct access.
-
-
-
-        Returns:
-
-            MuJoCo model object (mujoco.MjModel), or None if not loaded.
-
-        """
-
-        if self.sim is not None:
-            return self.sim.model
-
-        return None
-
-    @property
-    def is_initialized(self) -> bool:
-        """Check if the engine has a loaded environment and simulation."""
-
-        return self.env is not None and self.sim is not None
-
-    def load_from_path(self, path: str) -> None:
-        """Load environment by ID (passed as path)."""
-
-        if not MYOSUITE_AVAILABLE:
-            raise ImportError("MyoSuite not installed")
-
-        # Heuristic: if path ends with .xml, user might be confused.
-
-        # But for MyoSuite, we expect an Env ID.
-
-        # We'll treat the 'path' argument as the Env ID.
-
-        env_id = path.strip()
-
-        try:
-            env = gym.make(env_id)
-            env.reset()
-            sim = self._extract_sim_from_env(env)
-
-            self.env = env
-            self.sim = sim
-            self.env_id = env_id
-            self._dt = self.sim.model.opt.timestep
-
-        except (RuntimeError, TypeError, ValueError, AttributeError) as e:
-            self._reset_loaded_state()
-            logger.error("Failed to load MyoSuite environment '%s': %s", env_id, e)
-
-            raise
-
-    def load_from_string(self, content: str, extension: str | None = None) -> None:
-        """Loading from string is not supported for Gym environments."""
-
-        logger.error(
-            "MyoSuite does not support loading from string (requires Env ID registration)"
-        )
-
-        raise RuntimeError(
-            "MyoSuite does not support loading from string (requires Env ID registration)"
-        )
-
-    @precondition(lambda self: self.env is not None, "Environment must be loaded")
-    def reset(self) -> None:
-        """Reset environment."""
-
-        if self.env:
-            self.env.reset()
-            self._terminated = False
-
-    @precondition(
-        lambda self, dt=None: self.env is not None, "Environment must be loaded"
-    )
-    def step(self, dt: float | None = None) -> None:
-        """Step simulation.
-
-        Bridges Gym interface properly to pass set_control values as actions.
-        """
-        if not self.env:
-            return
-
-        if dt is not None and dt != getattr(self, "_dt", None):
-            logger.warning(
-                "Runtime timestep modification is unsafe in MuJoCo. Ignoring dt override."
-            )
-
-        if self._terminated:
-            logger.warning(
-                "step() called on a terminated MyoSuite environment; "
-                "call reset() before stepping again"
-            )
-            return
-
-        action = getattr(self, "_last_action", None)
-        if action is None:
-            # Fallback to zero action
-            action = self.env.action_space.sample() * 0.0
-
-        result = self.env.step(action)
-        # Gym step returns (obs, reward, terminated, truncated, info)
-        if len(result) >= 4:
-            _obs, _reward, terminated, truncated = result[:4]
-            self._terminated = bool(terminated or truncated)
-
-    @precondition(lambda self: self.is_initialized, "Engine must be initialized")
-    def forward(self) -> None:
-        """Compute forward dynamics."""
-
-        if self.sim:
-            self.sim.forward()
-
-    def get_state(self) -> tuple[np.ndarray, np.ndarray]:
-        """Get qpos, qvel."""
-
-        if not self.sim:
-            return np.array([]), np.array([])
-
-        return (np.array(self.sim.data.qpos[:]), np.array(self.sim.data.qvel[:]))
-
-    def set_state(self, q: np.ndarray, v: np.ndarray) -> None:
-        """Set joint positions and velocities on the simulation."""
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not self.sim:
-            return
-
-        # Ensure arrays are at least 1D to avoid len() errors on scalars
-
-        q = np.atleast_1d(q)
-
-        v = np.atleast_1d(v)
-
-        # MjSim (mujoco-py) usually allows direct assignment
-
-        # Handle both real arrays and mocked objects safely
-
-        try:
-            qpos = self.sim.data.qpos
-
-            qvel = self.sim.data.qvel
-
-            # Check if we can get lengths safely
-
-            if hasattr(qpos, "__len__") and hasattr(qvel, "__len__"):
-                if len(q) == len(qpos):
-                    self.sim.data.qpos[:] = q
-
-                if len(v) == len(qvel):
-                    self.sim.data.qvel[:] = v
-
-            else:
-                # Fallback for mocked objects or unusual cases
-
-                self.sim.data.qpos[:] = q
-
-                self.sim.data.qvel[:] = v
-
-        except (TypeError, AttributeError) as e:
-            # Handle mocked objects or other edge cases
-
-            logger.debug(f"Primary state assignment failed (may be mocked): {e}")
-
-            try:
-                self.sim.data.qpos[:] = q
-
-                self.sim.data.qvel[:] = v
-
-            except (RuntimeError, ValueError, OSError) as fallback_error:
-                # Log instead of silent pass - helps debugging test failures
-
-                logger.debug(f"Fallback state assignment failed: {fallback_error}")
-
-        try:
-            self.sim.forward()
-
-        except (RuntimeError, ValueError, OSError) as forward_error:
-            # Log instead of silent pass - helps debugging test failures
-
-            logger.debug(f"Forward dynamics failed (may be mocked): {forward_error}")
-
-    def set_control(self, u: np.ndarray) -> None:
-        """Set control (ctrl)."""
-        if not (u is not None):
-            raise ValueError("u must be provided")
-        if not (u is not None):
-            raise ValueError("u must be provided")
-        self._last_action = np.array(u, copy=True)
-
-        if not self.sim:
-            return
-
-        try:
-            ctrl = self.sim.data.ctrl
-
-            if hasattr(ctrl, "shape") and hasattr(ctrl, "__len__"):
-                if len(u) == ctrl.shape[0]:
-                    self.sim.data.ctrl[:] = u
-
-            else:
-                # Fallback for mocked objects
-
-                self.sim.data.ctrl[:] = u
-
-        except (TypeError, AttributeError) as e:
-            # Handle mocked objects or other edge cases
-
-            logger.debug(f"Primary control assignment failed (may be mocked): {e}")
-
-            try:
-                self.sim.data.ctrl[:] = u
-
-            except (RuntimeError, ValueError, OSError) as fallback_error:
-                # Log instead of silent pass - helps debugging test failures
-
-                logger.debug(f"Fallback control assignment failed: {fallback_error}")
-
-    def get_time(self) -> float:
-        """Return the current simulation time in seconds."""
-        if self.sim:
-            return float(self.sim.data.time)
-
-        return 0.0
-
-    @precondition(lambda self: self.is_initialized, "Engine must be initialized")
-    @postcondition(check_finite, "Mass matrix must contain finite values")
-    def compute_mass_matrix(self) -> np.ndarray:
-        """Compute the joint-space mass matrix via MuJoCo."""
-        if not self.sim:
-            return np.array([])
-
-        # Try generic mujoco approach assuming 'mujoco' lib is available
-
-        try:
-            import mujoco
-
-            # Handle both real MuJoCo objects and mocks
-
-            if hasattr(self.sim.model, "nv") and not isinstance(
-                self.sim.model.nv, type(lambda: None)
-            ):
-                nv = self.sim.model.nv
-
-            else:
-                # Fallback for mocked objects - assume small system
-
-                nv = 1
-
-            M = np.zeros((nv, nv))
-
-            # Try to call mj_fullM with proper error handling for mocks
-
-            try:
-                mujoco.mj_fullM(self.sim.model, M, self.sim.data.qM)
-
-            except TypeError:
-                # Handle mocked objects - return identity matrix
-
-                M = np.eye(nv)
-
-            return M
-
-        except ImportError as e:
-            logger.error("Failed to compute mass matrix: %s", e)
-
-            return np.array([])
-
-    @precondition(lambda self: self.is_initialized, "Engine must be initialized")
-    @postcondition(check_finite, "Bias forces must contain finite values")
-    def compute_bias_forces(self) -> np.ndarray:
-        """Return the Coriolis and gravity bias force vector."""
-        if self.sim:
-            return np.array(self.sim.data.qfrc_bias)
-
-        return np.array([])
-
-    def compute_gravity_forces(self) -> np.ndarray:
-        """Return gravity forces (not exposed in basic bindings)."""
-        # Not easily exposed separately in basic bindings without extra calc
-
-        return np.array([])
-
-    @precondition(lambda self, qacc: self.is_initialized, "Engine must be initialized")
-    @postcondition(check_finite, "Inverse dynamics torques must contain finite values")
-    def compute_inverse_dynamics(self, qacc: np.ndarray) -> np.ndarray:
-        """Compute inverse dynamics torques for the given acceleration."""
-        # Requires calling mj_inverse
-
-        if not (qacc is not None):
-            raise ValueError("qacc must be provided")
-        if not (qacc is not None):
-            raise ValueError("qacc must be provided")
-        if not self.sim:
-            return np.array([])
-
-        try:
-            import mujoco
-
-            self.sim.data.qacc[:] = qacc
-
-            mujoco.mj_inverse(self.sim.model, self.sim.data)
-
-            return np.array(self.sim.data.qfrc_inverse)
-
-        except ImportError as e:
-            logger.error("Failed to compute inverse dynamics: %s", e)
-
-            return np.array([])
-
-    def compute_jacobian(self, body_name: str) -> dict[str, np.ndarray] | None:
-        """Compute linear and angular Jacobian for a named body."""
-        if not (body_name is not None):
-            raise ValueError("body_name must be provided")
-        if not (body_name is not None):
-            raise ValueError("body_name must be provided")
-        if not self.sim:
-            return None
-
-        try:
-            import mujoco
-
-            body_id = mujoco.mj_name2id(
-                self.sim.model, mujoco.mjtObj.mjOBJ_BODY, body_name
-            )
-
-            if body_id == -1:
-                return None
-
-            jacp = np.zeros((3, self.sim.model.nv))
-
-            jacr = np.zeros((3, self.sim.model.nv))
-
-            mujoco.mj_jacBody(self.sim.model, self.sim.data, jacp, jacr, body_id)
-
-            return {"linear": jacp, "angular": jacr, "spatial": np.vstack([jacr, jacp])}
-
-        except ImportError as e:
-            logger.error("Failed to compute Jacobian for body '%s': %s", body_name, e)
-
-            return None
-
-    # -------- Section F: Drift-Control Decomposition --------
-
-    @precondition(lambda self: self.is_initialized, "Engine must be initialized")
-    @postcondition(check_finite, "Drift acceleration must contain finite values")
-    def compute_drift_acceleration(self) -> np.ndarray:
-        """Compute passive (drift) acceleration with zero muscle activations.
-
-
-
-        Section F Implementation: Uses MuJoCo forward dynamics with zero muscle activations
-
-        to isolate passive dynamics (gravity + Coriolis + constraints).
-
-
-
-        Returns:
-
-            q_ddot_drift: Drift acceleration vector (nv,) [rad/s² or m/s²]
-
-        """
-
-        if not self.sim:
-            logger.warning("Simulation not initialized")
-
-            return np.array([])
-
-        try:
-            # Save current activations/controls
-
-            ctrl_saved = self.sim.data.ctrl.copy()
-
-            # Set all muscle activations to zero
-
-            self.sim.data.ctrl[:] = 0.0
-
-            # Compute forward dynamics
-
-            # Use self.sim.forward() for compatibility with mujoco-py MjSim
-
-            self.sim.forward()
-
-            # Extract drift acceleration (explicit type for mypy)
-
-            a_drift: np.ndarray = np.array(self.sim.data.qacc)
-
-            # Restore original controls
-
-            self.sim.data.ctrl[:] = ctrl_saved
-
-            self.sim.forward()
-
-            return a_drift
-
-        except (ValueError, TypeError, RuntimeError) as e:
-            logger.error(f"Failed to compute drift acceleration: {e}")
-
-            return np.array([])
-
-    @precondition(lambda self, tau: self.is_initialized, "Engine must be initialized")
-    @postcondition(check_finite, "Control acceleration must contain finite values")
-    def compute_control_acceleration(self, tau: np.ndarray) -> np.ndarray:
-        """Compute control-attributed acceleration from muscle activations.
-
-
-
-        Section F Implementation: Computes M(q)^-1 * tau to isolate control component.
-
-
-
-        Args:
-
-            tau: Applied generalized forces (nv,) [N·m or N]
-
-
-
-        Returns:
-
-            q_ddot_control: Control acceleration vector (nv,) [rad/s² or m/s²]
-
-        """
-
-        if not (tau is not None):
-            raise ValueError("tau must be provided")
-        if not (tau is not None):
-            raise ValueError("tau must be provided")
-        if not self.sim:
-            logger.warning("Simulation not initialized")
-
-            return np.array([])
-
-        try:
-            # Get mass matrix
-
-            M = self.compute_mass_matrix()
-
-            if M.size == 0:
-                return np.array([])
-
-            # Control component: M^-1 * tau
-
-            a_control = np.linalg.solve(M, tau)
-
-            return a_control
-
-        except (ValueError, TypeError, RuntimeError) as e:
-            logger.error(f"Failed to compute control acceleration: {e}")
-
-            return np.zeros_like(tau)
-
-    # -------- Section K: MyoSuite Muscle Integration --------
-
-    def get_muscle_analyzer(self) -> Any | None:
-        """Get muscle analyzer for biomechanical analysis.
-
-
-
-        Section K: Provides access to muscle-specific analysis capabilities.
-
-
-
-        Returns:
-
-            MyoSuiteMuscleAnalyzer instance or None if sim not ready
-
-        """
-
-        if not self.sim:
-            logger.warning("Cannot create muscle analyzer - simulation not initialized")
-
-            return None
-
-        try:
-            from .muscle_analysis import MyoSuiteMuscleAnalyzer
-
-            return MyoSuiteMuscleAnalyzer(self.sim)
-
-        except ImportError as e:
-            logger.error(f"Failed to import muscle analyzer: {e}")
-
-            return None
-
-    def create_grip_model(self) -> Any | None:
-        """Create grip modeling interface.
-
-
-
-        Section K1: Provides activation-driven grip force analysis.
-
-
-
-        Returns:
-
-            MyoSuiteGripModel instance or None if analyzer not ready
-
-        """
-
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            logger.warning("Cannot create grip model - muscle analyzer not available")
-
-            return None
-
-        try:
-            from .muscle_analysis import MyoSuiteGripModel
-
-            return MyoSuiteGripModel(self.sim, analyzer)
-
-        except ImportError as e:
-            logger.error(f"Failed to import grip model: {e}")
-
-            return None
-
-    def set_muscle_activations(self, activations: dict[str, float]) -> None:
-        """Set muscle activation levels by name.
-
-
-
-        Section K: Neural control interface for muscle-driven simulation.
-
-
-
-        Args:
-
-            activations: Dictionary mapping muscle names to activation [0-1]
-
-        """
-
-        if not (activations is not None):
-            raise ValueError("activations must be provided")
-        if not (activations is not None):
-            raise ValueError("activations must be provided")
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            logger.warning("Cannot set activations - muscle analyzer unavailable")
-
-            return
-
-        # Map muscle names to actuator indices
-
-        for muscle_name, activation in activations.items():
-            try:
-                idx = analyzer.muscle_names.index(muscle_name)
-
-                actuator_id = analyzer.muscle_actuator_ids[idx]
-
-                # Clip to valid range
-
-                activation_clamped = max(0.0, min(1.0, activation))
-
-                # Set control
-
-                try:
-                    ctrl = self.sim.data.ctrl
-
-                    if hasattr(ctrl, "__len__") and actuator_id < len(ctrl):
-                        self.sim.data.ctrl[actuator_id] = activation_clamped
-
-                    elif hasattr(ctrl, "__setitem__"):
-                        # Fallback for mocked objects
-
-                        self.sim.data.ctrl[actuator_id] = activation_clamped
-
-                except (TypeError, AttributeError, IndexError):
-                    # Handle mocked objects or other edge cases
-
-                    pass
-
-            except ValueError:
-                logger.warning(f"Muscle '{muscle_name}' not found")
-
-            except (RuntimeError, OSError) as e:
-                logger.error(f"Failed to set activation for '{muscle_name}': {e}")
-
-    def compute_muscle_induced_accelerations(self) -> dict[str, np.ndarray]:
-        """Compute acceleration contributions from each muscle.
-
-
-
-        Section K Requirement: Muscle contribution to joint accelerations.
-
-
-
-        Returns:
-
-            Dictionary mapping muscle names to induced accelerations [rad/s²]
-
-        """
-
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            return {}
-
-        return dict(analyzer.compute_muscle_induced_accelerations())
-
-    def analyze_muscle_contributions(self) -> Any | None:
-        """Full muscle contribution analysis.
-
-
-
-        Section K Requirement: Comprehensive muscle reports (forces, moments, power).
-
-
-
-        Returns:
-
-            MyoSuiteMuscleAnalysis object with all muscle metrics
-
-        """
-
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            logger.warning("Cannot analyze muscles - analyzer not available")
-
-            return None
-
-        return analyzer.analyze_all()
-
-    def get_muscle_state(self) -> Any | None:
-        """Get current muscle state.
-
-
-
-        Section K: Muscle state for monitoring and control.
-
-
-
-        Returns:
-
-            MyoSuiteMuscleState with activations, forces, lengths, velocities
-
-        """
-
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            return None
-
-        from .muscle_analysis import MyoSuiteMuscleState
-
-        return MyoSuiteMuscleState(
-            muscle_names=analyzer.muscle_names,
-            activations=analyzer.get_muscle_activations(),
-            forces=analyzer.get_muscle_forces(),
-            lengths=analyzer.get_muscle_lengths(),
-            velocities=analyzer.get_muscle_velocities(),
-        )
-
-    def compute_ztcf(self, q: np.ndarray, v: np.ndarray) -> np.ndarray:
-        """Zero-Torque Counterfactual (ZTCF) - Guideline G1.
-
-
-
-        Compute acceleration with applied torques set to zero, preserving current state.
-
-        This isolates drift (gravity + Coriolis + constraints) from control effects.
-
-
-
-        Args:
-
-            q: Joint positions (n_v,)
-
-            v: Joint velocities (n_v,)
-
-
-
-        Returns:
-
-            q̈_ZTCF: Acceleration under zero applied torque (n_v,)
-
-        """
-
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not self.sim:
-            return np.array([])
-
-        try:
-            # Save current state
-
-            q_saved, v_saved = self.get_state()
-
-            ctrl_saved = self.sim.data.ctrl.copy()
-
-            # Set desired state
-
-            self.set_state(q, v)
-
-            # Set zero control
-
-            self.sim.data.ctrl[:] = 0.0
-
-            # Compute forward dynamics
-
-            # Use self.sim.forward() to support MjSim
-
-            self.sim.forward()
-
-            a_ztcf = np.array(self.sim.data.qacc)
-
-            # Restore state and control
-
-            self.sim.data.ctrl[:] = ctrl_saved
-
-            self.set_state(q_saved, v_saved)
-
-            return a_ztcf
-
-        except (ValueError, TypeError, RuntimeError) as e:
-            logger.error(f"Failed to compute ZTCF: {e}")
-            return np.array([])
-
-    def compute_zvcf(self, q: np.ndarray) -> np.ndarray:
-        """Zero-Velocity Counterfactual (ZVCF) - Guideline G2.
-
-
-
-        Compute acceleration with joint velocities set to zero, preserving configuration
-
-        and controls. This isolates configuration-dependent effects (gravity, constraints).
-
-
-
-        Args:
-
-            q: Joint positions (n_v,)
-
-
-
-        Returns:
-
-            q̈_ZVCF: Acceleration with v=0 (n_v,)
-
-        """
-
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not (q is not None):
-            raise ValueError("q must be provided")
-        if not self.sim:
-            return np.array([])
-
-        try:
-            # Save current state
-
-            q_saved, v_saved = self.get_state()
-
-            # Set state with zero velocity
-
-            try:
-                # Fallback for scalar or mocked objects
-                n_v = len(v_saved) if hasattr(v_saved, "__len__") else 1
-
-            except TypeError:
-                n_v = 1
-
-            self.set_state(q, np.zeros(n_v))
-
-            # Controls are preserved in data.ctrl automatically unless we change them
-
-            # Compute forward dynamics
-
-            self.sim.forward()
-
-            a_zvcf = np.array(self.sim.data.qacc)
-
-            # Restore state
-
-            self.set_state(q_saved, v_saved)
-
-            return a_zvcf
-
-        except (ValueError, TypeError, RuntimeError) as e:
-            logger.error(f"Failed to compute ZVCF: {e}")
-
-            return np.array([])
-
-    def get_acceleration(self) -> np.ndarray:
-        """Get current acceleration vector.
-
-
-
-        Returns:
-
-            Current joint accelerations (nv,) [rad/s² or m/s²]
-
-        """
-
-        if not self.sim:
-            return np.array([])
-
-        return np.array(self.sim.data.qacc)
-
-    def get_muscle_names(self) -> list[str]:
-        """Get list of muscle names in the model.
-
-
-
-        Returns:
-
-            List of muscle actuator names
-
-        """
-
-        analyzer = self.get_muscle_analyzer()
-
-        if analyzer is None:
-            return []
-
-        return list(analyzer.muscle_names)
+        EngineInitMixin.__init__(self)
 
 
 def main() -> None:
